modified demo class to read input from file
had a bit of trouble with reading the input (can't believe I forgot about the remaining newline, thanks Prof.) so I added a second Policy.text file to test a few things, however, that was after I accidentally modified the PolicyInformation.txt file, and was able to fix that. Either way, the mess is cleared now! And I'm able to submit the project! yay!