finalized policy holder class
had to revert changes to the smoke status variable and methods. couldn't get anything to work properly with smoke status as a String, due to an error message saying the variable and subsequent methods were null. i had no idea how to fix that, so i reverted the variable to a Boolean and reset the methods. then everything worked fine

i also added the toString method